+++ app/Self/component/Shared.jsx
@@ -1,12 +1,18 @@
 import React from 'react';
 import {
   Route,
-  Link
+  Link,
+  withRouter
 } from 'react-router-dom';
+import {connect} from "react-redux";
+import NailShared from './NailShared.jsx';
+import Threads from './Threads.jsx';
 import CreateShare from '../../Component/CreateShare.jsx';
 import Unit from '../../Component/Unit.jsx';
+//ModalBox used some unstable method, considering updating some day.
+import ModalBox from '../../Component/ModalBox.jsx';
 
-export default class Shared extends React.Component {
+class Shared extends React.Component {
   constructor(props){
     super(props);
     this.state = {
@@ -44,24 +50,6 @@ export default class Shared extends React.Component {
         left: '0',
         boxSizing: 'border-box',
         padding: '2vh 0 0 0'
-      },
-      selfCom_Shared_nails_div_: {
-        display: 'inline-block',
-        width: '32%',
-        height: '32vh',
-        position: 'relative',
-        boxSizing: 'border-box',
-        margin: '0 1% 2vh 0',
-        overflow: 'hidden',
-        cursor: 'pointer'
-      },
-      selfCom_Shared_nails_div_img: {
-        maxWidth: '150%',
-        maxHeight: '150%',
-        position: 'absolute',
-        top: '50%',
-        left: '50%',
-        transform: 'translate(-50%,-50%)'
       }
     }
   }
@@ -107,21 +95,11 @@ export default class Shared extends React.Component {
     let shares = self.state.unitsList.map(function(dataKey, index){
       let dataValue = self.state.unitsBasicSet[dataKey];
       return(
-        <Link
+        <NailShared
+          {...self.props}
           key={'key_Shared_nails_'+index}
-          to={{
-            pathname: self.props.match.url+"/units/"+dataKey,
-            state: {from: self.props.location}
-          }}>
-
-          <div
-            unitname={dataKey}
-            style={self.style.selfCom_Shared_nails_div_}>
-            <img
-              src={'/router/img/'+dataValue.pic_layer0+'?type=thumb'}
-              style={self.style.selfCom_Shared_nails_div_img}/>
-          </div>
-        </Link>
+          sharedId={dataKey}
+          unitBasic={dataValue}/>
       )
     })
 
@@ -134,7 +112,6 @@ export default class Shared extends React.Component {
             style={this.style.selfCom_Shared_top_CreateShare_}>
             <img src="/images/vacancy.png" style={{width: '100%', height: '100%'}}/>
             <CreateShare
-              userBasic={this.props.userBasic}
               _submit_Share_New={this._submit_Share_New}/>
           </div>
         </div>
@@ -142,8 +119,23 @@ export default class Shared extends React.Component {
           style={this.style.selfCom_Shared_nails_}>
           {shares}
         </div>
+        <ModalBox containerId="root">
+          <Route path={this.props.match.path+"/:sharedId/threads"} render={(props)=> <Threads {...props} unitBasic={this.state.unitsBasicSet[props.match.params.sharedId]} _refer_leaveSelf={this.props._refer_leaveSelf}/>}/>
+        </ModalBox>
         <Route path={this.props.match.path+"/units/:id"} render={(props)=> <Unit {...props} _construct_UnitInit={this._construct_UnitInit} _refer_von_unit={this.props._refer_leaveSelf}/>}/>
       </div>
     )
   }
 }
+
+const mapStateToProps = (state)=>{
+  return {
+    userInfo: state.userInfo,
+    unitCurrent: state.unitCurrent
+  }
+}
+
+export default withRouter(connect(
+  mapStateToProps,
+  null
+)(Shared));
