+++ app/Within/component/LtdNav.jsx
@@ -72,7 +72,7 @@ class LtdNav extends React.Component {
   }
 
   _submit_Share_New(dataObj){
-    window.location.assign('/user/front');
+    window.location.assign('/user/front/actions/shareds');
   }
 
   _handleClick_selfEntrance(event){
