+++ app/redux/constants/general.js
@@ -1,3 +1,5 @@
 export const MOUNT_USERINFO = 'MOUNT_USERINFO';
 
+export const UNIT_MOUNT_UNITCURRENT = 'UNIT_MOUNT_UNITCURRENT';
+
 export const SET_VISIBILITY_FILTER = 'SET_VISIBILITY_FILTER'; //from original example
