+++ app/redux/reducers/selfCover.js
@@ -1,9 +1,11 @@
 import {
-  MOUNT_USERINFO
+  MOUNT_USERINFO,
+  UNIT_MOUNT_UNITCURRENT
 } from '../constants/general.js';
 
 const initialGeneral = {
-  userInfo: {}
+  userInfo: {},
+  unitCurrent: {unitId:"", identity: ""}
 }
 
 function pageSelfCover(state = initialGeneral, action){
@@ -11,6 +13,11 @@ function pageSelfCover(state = initialGeneral, action){
     case MOUNT_USERINFO:
       return Object.assign({}, state, {
         userInfo: action.userInfo
+      })
+      break;
+    case UNIT_MOUNT_UNITCURRENT:
+      return Object.assign({}, state, {
+        unitCurrent: action.unitCurrent
       })
       break;
     default:
