+++ app/redux/reducers/selfFront.js
@@ -1,9 +1,11 @@
 import {
-  MOUNT_USERINFO
+  MOUNT_USERINFO,
+  UNIT_MOUNT_UNITCURRENT
 } from '../constants/general.js';
 
 const initialGeneral = {
-  userInfo: {}
+  userInfo: {},
+  unitCurrent: {unitId:"", identity: ""}
 }
 
 function pageSelfFront(state = initialGeneral, action){
@@ -11,6 +13,11 @@ function pageSelfFront(state = initialGeneral, action){
     case MOUNT_USERINFO:
       return Object.assign({}, state, {
         userInfo: action.userInfo
+      })
+      break;
+    case UNIT_MOUNT_UNITCURRENT:
+      return Object.assign({}, state, {
+        unitCurrent: action.unitCurrent
       })
       break;
     default:
