+++ docs/routes.md
@@ -1,0 +1,45 @@
+# routes
+
+## /
+>html_Within.html
+
+#### /
+
+#### /cosmic
+* /units/:id/related
+  * /:id/origin
+  * /:id/responses
+* /people/:id
+* /nouns/:id
+* /series
+* /series/:id
+
+## /user
+>html_SelfFront.html<br> html_SelfCover.html
+
+#### /overview
+  * /appear
+  * /period
+
+#### /front
+  * /actions
+    * /shareds
+      * /:sharedId/threads
+    * /events
+  * /mutuals
+    * /dialogues
+  * /wall
+    * /inspireds
+    * /broads
+  * /series
+
+#### /collateral
+  * /statics
+
+#### /...
+* /units/:id
+* /:unitId
+
+/login
+
+/
