+++ src/units/main.js
@@ -3,6 +3,7 @@ const main = express.Router();
 
 const basicExecutive = require('./basic.js');
 const responsesExecutive = require('./responses.js');
+const vanilla = require('./vanilla.js');
 
 main.param("id", (req, res, next, id)=>{
   req.reqUnitId = id;
@@ -11,5 +12,7 @@ main.param("id", (req, res, next, id)=>{
 
 main.use('/:id/basic', basicExecutive)
 main.use('/:id/responses', responsesExecutive)
+
+main.use('/:id', vanilla)
 
 module.exports = main;
