+++ src/utils/queryIndicators.js
@@ -1,0 +1,5 @@
+const DIALOGUES_LATEST = "DIALOGUES_LATEST";
+
+module.exports = {
+  DIALOGUES_LATEST: DIALOGUES_LATEST
+};
